Merging into the new functional form. The new formulas allows to specify alternative deterministic functions for the SDE of trait evolution.

+++ src/compete_utils.jl
@@ -66,7 +66,7 @@ logdhcau1(x::Float64) =
 Generate one sample from a random **Exponential** variable
 with mean `λ`. 
 """
-rexp(λ::Float64) = @fastmath log(rand()) * -(1/λ)
+rexp(λ::Float64) = @fastmath (log(rand()) * -(1.0/λ))::Float64
